% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_functions.R
\name{add_go_heat}
\alias{add_go_heat}
\title{Annotate clusters with top GO term}
\usage{
add_go_heat(marker_lab, enrich_list, top_go_name = 1)
}
\arguments{
\item{marker_lab}{Integer or character vector. Cluster assignments for each gene.}

\item{enrich_list}{List of data frames, each containing GO enrichment results for a cluster.}

\item{top_go_name}{Integer. Which GO term to use (row index in enrichment result). Default is 1.}
}
\value{
A vector of GO term descriptions, same length as marker_lab.
}
\description{
For each cluster, assigns the top GO term description to the marker label vector.
}
\examples{
\dontrun{
  add_go_heat(marker_lab, enrich_list, top_go_name = 1)
}
}
