% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaPE.R
\name{AlphaPE}
\alias{AlphaPE}
\title{Alpha from the projection score calculation. 
Calculates the fraction of the variance explained by the chosen principal components and moderates it by taking the square root.}
\usage{
AlphaPE(DataMatrix, NPCs = 3)
}
\arguments{
\item{DataMatrix.}{Dataset in matrix form.}

\item{NPCs.}{Number of principal components that are considered informative for the score.}
}
\value{
alphaFinal. Fraction of variance explained by the principal components for the projection score calculation.
}
\description{
Alpha from the projection score calculation. 
Calculates the fraction of the variance explained by the chosen principal components and moderates it by taking the square root.
}
\examples{
AlphaPE(DataMatrix, NPCs = 3)

}
