% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltasScoreHeatmap.R
\name{AtlasScoreHeatmap}
\alias{AtlasScoreHeatmap}
\title{Title}
\usage{
AtlasScoreHeatmap(ExprMat, pure_markers, Csplit, Rsplit, colnamesS = F)
}
\arguments{
\item{colnamesS}{}
}
\value{
A heatmap with PC score
}
\description{
Title
}
\examples{
AtlasScoreHeatmap()
}
