% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootAlpha.R
\name{BootAlpha}
\alias{BootAlpha}
\title{Generates bootstrap samples of alpha for projection score calculation.
Takes a data matrix and resamples it to obtain a distribution for the fraction of variance explained by a given number of principal components.}
\usage{
BootAlpha(DataMatrix, NPCs = 3, nboot = 100)
}
\arguments{
\item{nboot}{}
}
\value{
Fraction of variance for projection score bootstrapped average
}
\description{
Generates bootstrap samples of alpha for projection score calculation.
Takes a data matrix and resamples it to obtain a distribution for the fraction of variance explained by a given number of principal components.
}
\examples{
AlphaPE()

}
