% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltasScoreHeatmap.R
\name{BulkPlots}
\alias{BulkPlots}
\title{Title}
\usage{
BulkPlots(ExprMat,Prefix,Contrastes,annotations,Treatment,species,logfold)
}
\arguments{
\item{colnamesS}{}
}
\value{
A heatmap with PC score
}
\description{
Title
}
\examples{
BulkPlots()
}
