% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterProfilerOnthologies.R
\name{ClusterProfilerOnthologies}
\alias{ClusterProfilerOnthologies}
\title{Title}
\usage{
ClusterProfilerOnthologies(stats, species, Prefix, contraste)
}
\arguments{
\item{contraste}{}
}
\value{
Plots
}
\description{
Title
}
\examples{
ClusterProfilerOnthologies()
}
