% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorPalettes.R
\name{ColorPalettes}
\alias{ColorPalettes}
\title{Get a Color Palette}
\usage{
ColorPalettes(palette = "TreatmentPalette")
}
\arguments{
\item{palette}{Character. Name of the palette to return. Options:
"TreatmentPalette", "ClusterPaletteCoarse", "CellcylcePal".
Default is "TreatmentPalette".}
}
\value{
A character vector of color hex codes.
}
\description{
Returns one of the predefined color palettes used across the package.
}
\examples{
ColorPalettes("TreatmentPalette")
ColorPalettes("ClusterPaletteCoarse")
ColorPalettes("CellcylcePal")
}
