% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{compute_mean}
\alias{compute_mean}
\title{calculate average}
\usage{
compute_mean(x)
}
\arguments{
\item{x}{A vector of numerical values}
}
\value{
A number
}
\description{
calculate average
}
\examples{
compute_mean(x=c(1,1,0,0))
}
