% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMouseGeneList.R
\name{convertMouseGeneList}
\alias{convertMouseGeneList}
\title{Title}
\usage{
convertMouseGeneList(x)
}
\arguments{
\item{x}{}
}
\value{
Unique gene
}
\description{
Title
}
\examples{
convertMouseGeneList()
}
