% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GostWrap.R
\name{GostWrap}
\alias{GostWrap}
\title{Wrapper for gostplot}
\usage{
GostWrap(signature, species = "mmusculus")
}
\arguments{
\item{species}{}
}
\value{
Wrapper for gostplot
}
\description{
Wrapper for gostplot
}
\examples{
GostWrap()
}
