% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsScore.R
\name{gsScore}
\alias{gsScore}
\title{Scores samples with a genesets}
\usage{
gsScore(ExprMat, GeneList, summarizationFunction = "PC")
}
\arguments{
\item{GeneList}{}
}
\value{
PC score
}
\description{
Scores samples with a genesets
}
\examples{
gsScore( )
}
