% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_qc_parameter.R
\name{highlight_qc_parameter}
\alias{highlight_qc_parameter}
\title{Highlights the cells to be removed/preserved after hard filtering on a particular quality control parameter}
\usage{
highlight_qc_parameter(seurat_object, qc_parameter, threshold)
}
\arguments{
\item{seurat_object.}{Seurat object with precomputed UMAP}

\item{qc_parameter.}{QC parameter on the metadata to color by.}

\item{threshold.}{Filter value.}
}
\value{
UMAP with hightlighted cells to be abova and below a QC threshold.
}
\description{
Highlights the cells to be removed/preserved after hard filtering on a particular quality control parameter
}
\examples{
highlight_qc_parameter()

}
