% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionPlotIrtx.R
\name{InteractionPlotIrtx}
\alias{InteractionPlotIrtx}
\title{Title}
\usage{
InteractionPlotIrtx(gensmbl, x.factor, trace.factor, ExprMat, cpalette)
}
\arguments{
\item{cpalette}{}
}
\value{
save ggplot
}
\description{
Title
}
\examples{
InteractionPlotIrtx( )
}
