% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadGseaDb.R
\name{LoadGSEADB}
\alias{LoadGSEADB}
\title{Loads gene sets for gene onthology analysis and gene set enrichment anlysis.}
\usage{
LoadGSEADB(species, date = "20251010")
}
\arguments{
\item{species.}{Dataset reference species either "Mus musculus" or "Homo sapiens".}

\item{date.}{To be downloaded rWikipPathWays database date.}
}
\value{
A list of gene set databases.
}
\description{
Loads gene sets for gene onthology analysis and gene set enrichment anlysis.
}
\examples{
LoadGSEADB(species = "Mus musculus" ,date="20241110")
}
