% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_seurat_objects.R
\name{merge_seurat_objects}
\alias{merge_seurat_objects}
\title{Merge Multiple Seurat Objects}
\usage{
merge_seurat_objects(seurats)
}
\arguments{
\item{seurats}{A character vector of file paths to the Seurat objects to be merged.}
}
\value{
A merged Seurat object with joined layers.
}
\description{
This function reads multiple Seurat objects from a list of file paths, merges them into a single Seurat object, and joins the layers.
}
\examples{
\dontrun{
seurats <- c("path/to/seurat1.rds", "path/to/seurat2.rds")
merged_seurat <- merge_seurat_objects(seurats)
}
}
