% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NthMostVariable.R
\name{NthMostVariableFeatures}
\alias{NthMostVariableFeatures}
\title{Title}
\usage{
NthMostVariableFeatures(mat, nvar)
}
\arguments{
\item{nvar}{}
}
\value{
Matrix
}
\description{
Title
}
\examples{
NthMostVariableFeatures( )
}
