% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionScore.R
\name{ProjectionScore}
\alias{ProjectionScore}
\title{Title}
\usage{
ProjectionScore(DataMatrix, NPCs = 3, nboot = 100)
}
\arguments{
\item{nboot}{}
}
\value{
Fraction of variance for projection score bootstrapped average
}
\description{
Title
}
\examples{
ProjectionScore(DataMatrix)

}
