% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratClustering.R
\name{SeuratClustering}
\alias{SeuratClustering}
\title{Remove Clustering Columns and Find Clusters}
\usage{
SeuratClustering(seu, prefix, resolutions)
}
\arguments{
\item{seu}{A Seurat object.}

\item{prefix}{A string encoding the prefix of the columns to be removed from the metadata.}

\item{resolutions}{A numeric vector of resolutions at which to find clusters.}
}
\value{
A Seurat object with updated clustering information.
}
\description{
This function removes columns from the metadata of a Seurat object that start with a specified prefix,
and then finds clusters at specified resolutions.
}
\examples{
seu <- SeuratClustering(seu, "res_", c(0.1, 0.2, 0.3, 0.4, 0.5))
}
