% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratFindTopMarkers.R
\name{SeuratFindTopMarkers}
\alias{SeuratFindTopMarkers}
\title{Process Seurat Object and Find Top Markers}
\usage{
SeuratFindTopMarkers(seu)
}
\value{
A data frame of top markers for each cluster.
}
\description{
This function processes a Seurat object by normalizing and scaling the data, 
and then finds the top markers for each cluster.
}
\examples{
\dontrun{
  # Assuming 'seu' is your Seurat object
  top_markers <- SeuratFindTopMarkers(seu)
}
}
