% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratTalusPlot.R
\name{SeuratTalusPlot}
\alias{SeuratTalusPlot}
\title{Plot PCA Standard Deviations from a Seurat Object}
\usage{
SeuratTalusPlot(seu, ndims = 30)
}
\arguments{
\item{seu}{A Seurat object containing PCA reduction.}

\item{ndims}{An integer specifying the number of dimensions to plot. Default is 30.}
}
\value{
A plot of the PCA standard deviations.
}
\description{
This function takes a Seurat object and plots the standard deviations of the PCA reduction to aid in selecting the optimal number of principal components.
}
\examples{
\dontrun{
  SeuratTalusPlot(seurat_object, 30)
}
}
