% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TalusPlot.R
\name{TalusPlot}
\alias{TalusPlot}
\title{Title}
\usage{
TalusPlot(PCAToolsObject, cutoff)
}
\arguments{
\item{cutoff}{}
}
\value{
The talus plot of a PCAtol pca object
}
\description{
Title
}
\examples{
TalusPlot()

}
