% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThetaProjectionScore.R
\name{ThetaProjectionScore}
\alias{ThetaProjectionScore}
\title{Sweep theta values for projection score}
\usage{
ThetaProjectionScore(
  DataMatrix,
  NPCs = 3,
  nboot = 100,
  thetas = c(0.25, 0.5, 0.75, 1)
)
}
\arguments{
\item{thetas}{}
}
\value{
A data.frame with columns Thetas and ProjectionScores.
}
\description{
Sweep theta values for projection score
}
\examples{
ps.df = ThetaProjectionScore(DataMatrix,NPCs =  3, nboot = 100, thetas =c(0.005,0.01,0.02,0.05,0.08,0.1,0.12,0.15,0.17,0.18,0.2,0.23,0.25,0.3,0.5,0.75,1))

}
