% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoWrap.R
\name{VolcanoWrap}
\alias{VolcanoWrap}
\title{Title}
\usage{
VolcanoWrap(signature, Prefix, contraste)
}
\arguments{
\item{contraste}{}
}
\value{
Volcano plot
}
\description{
Title
}
\examples{
VolcanoWrap( )
}
