% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_functions.R
\name{add_go_heat}
\alias{add_go_heat}
\title{Title}
\usage{
add_go_heat(marker_lab, enrich_list, top_go_name = 1)
}
\arguments{
\item{top_go_name}{}
}
\value{
GO sescription
}
\description{
Title
}
\examples{
add_go_heat( )
}
