% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkPlot.R
\name{BulkPlot}
\alias{BulkPlot}
\title{Title}
\usage{
BulkPlot(signature, Prefix, contraste, species, Foldstatistic = "logFC")
}
\arguments{
\item{Foldstatistic}{}
}
\value{
Master wrapper for multiple functions
}
\description{
Title
}
\examples{
BulkPlot()
}
