
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltasScoreHeatmap.R
\name{BulkPlot}
\alias{BulkPlot}
\title{Title}
\usage{
BulkPlot(signature,Prefix,contraste,species)
}
\arguments{
\item{colnamesS}{}
}
\value{
A heatmap with PC score
}
\description{
Title
}
\examples{
BulkPlot()
}
