% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkPlots.R
\name{BulkPlots}
\alias{BulkPlots}
\title{Loops BulkPlot through a set of limma outputs.}
\usage{
BulkPlots(
  ExprMat,
  Prefix,
  Contrastes,
  annotations,
  Treatment,
  species,
  logfold
)
}
\arguments{
\item{ExprMat}{Normalized expression matrix}

\item{Prefix}{String to append to the contrasts}

\item{logfold}{}
}
\value{
Master wrapper for multiple functions
}
\description{
Loops BulkPlot through a set of limma outputs.
}
\examples{
BulkPlots()
}
