% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkPlots.R
\name{BulkPlots}
\alias{BulkPlots}
\title{Loops BulkPlot through a set of limma outputs}
\usage{
BulkPlots(
  ExprMat,
  Prefix,
  Contrastes,
  annotations,
  Treatment,
  species,
  logfold
)
}
\arguments{
\item{ExprMat}{Normalized expression matrix.}

\item{Prefix}{String to append to the contrasts.}

\item{Contrastes}{A vector of contrast names.}

\item{annotations}{Annotations for the genes (not used in the function but can be included for completeness).}

\item{Treatment}{Treatment conditions (not used in the function but can be included for completeness).}

\item{species}{A character string indicating the species (e.g., "human", "mouse").}

\item{logfold}{Log fold change threshold for differential expression.}
}
\value{
This function does not return a value but generates several output files and plots.
}
\description{
This function loops through a set of contrasts from limma outputs and generates various plots and files for differential gene expression analysis.
}
\examples{
# Example usage:
BulkPlots(ExprMat, Prefix = "Sample", Contrastes = c("ConditionA_vs_ConditionB", "ConditionC_vs_ConditionD"), annotations, Treatment, species = "human", logfold = 1)
}
