% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkPlots.R
\name{BulkPlots}
\alias{BulkPlots}
\title{Title}
\usage{
BulkPlots(
  ExprMat,
  Prefix,
  Contrastes,
  annotations,
  Treatment,
  species,
  logfold
)
}
\arguments{
\item{logfold}{}
}
\value{
Master wrapper for multiple functions
}
\description{
Title
}
\examples{
BulkPlots()
}
