% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterProfilerPlots.R
\name{ClusterProfilerPlots}
\alias{ClusterProfilerPlots}
\title{Title}
\usage{
ClusterProfilerPlots(ewp2, Prefix, contraste, DataBase, geneList)
}
\arguments{
\item{geneList}{}
}
\value{
GSEA plot
}
\description{
Title
}
\examples{
ClusterProfilerPlots()
}
