% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorPalettes.R
\name{ColorPalettes}
\alias{ColorPalettes}
\title{Get a Color Palette}
\usage{
ColorPalettes(palette = "palette1")
}
\arguments{
\item{palette}{A character string specifying the palette to use. Options are "palette1", "palette2", and "palette3".}
}
\value{
A vector of color hex codes.
}
\description{
This function returns a specified color palette.
}
\examples{
ColorPalettes("palette1")
}
