% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMouseGeneList.R
\name{convertMouseGeneList}
\alias{convertMouseGeneList}
\title{Convert Mouse Gene Symbols to Human Gene Symbols}
\usage{
convertMouseGeneList(x)
}
\arguments{
\item{x}{A character vector of mouse gene symbols (e.g., "mgi_symbol").}
}
\value{
A character vector of unique human gene symbols corresponding to the input mouse gene symbols.
}
\description{
This function converts a list of mouse gene symbols to their corresponding human gene symbols using the `biomaRt` package. It queries the Ensembl database to perform the mapping.
}
\examples{
# Example usage:
# Assuming `mouse_genes` is a character vector of mouse gene symbols:
mouse_genes <- c("Tnf", "Il6", "Cd4")
human_genes <- convertMouseGeneList(mouse_genes)
print(human_genes)

}
