% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project_structure.R
\name{create_project_structure}
\alias{create_project_structure}
\title{Create a Standard Project Directory Structure}
\usage{
create_project_structure(top_dir = "NewProject")
}
\arguments{
\item{top_dir}{Character. The name of the top-level directory where the structure will be created.}
}
\value{
Creates directories on disk. Returns a message upon successful creation.
}
\description{
This function creates a predefined directory structure for a computational biology project
under a user-specified top-level directory. The structure includes folders for data,
source code, QC, clustering, analysis, figures, slides, and paper drafts.
}
\examples{
create_project_structure("MyProject")
}
