% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichOut.R
\name{enrich_out}
\alias{enrich_out}
\title{Title}
\usage{
enrich_out(
  expmat,
  nk,
  scale = TRUE,
  organism,
  top_go = 3,
  Method = c("GO", "Wiki", "Kegg", "Custom"),
  cust_db = NULL
)
}
\arguments{
\item{cust_db}{}
}
\value{
Objects to GOHeatmap
}
\description{
Title
}
\examples{
enrich_out()
}
