% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterLowlyExpressed.R
\name{FilterLowlyExpressed}
\alias{FilterLowlyExpressed}
\title{Title}
\usage{
FilterLowlyExpressed(y.all, nsamples = 3, nreads = 15)
}
\arguments{
\item{nsamples}{}
}
\value{
Projection Score Plot
}
\description{
Title
}
\examples{
FilterLowlyExpressed(y.all,nsamples = 3)

}
