% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_functions.R
\name{go_enrich_save}
\alias{go_enrich_save}
\title{Title}
\usage{
go_enrich_save(enrich_list, filename)
}
\arguments{
\item{filename}{}
}
\value{
Dowload files
}
\description{
Title
}
\examples{
go_enrich_save( )
}
