% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOheatmap.R
\name{GOHeatMap}
\alias{GOHeatMap}
\title{Title}
\usage{
GOHeatMap(
  scaled_mat,
  Gomet,
  nk_Go,
  organismRef,
  nHL = 5,
  Csplit,
  gf = 5,
  title = "  ",
  clusCol = T,
  top_annotation,
  cust_db = NULL,
  hl = T
)
}
\arguments{
\item{hl}{}
}
\value{
GO Heatmap
}
\description{
Title
}
\examples{
GOHeatMap( )
}
