% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAprepare.R
\name{GSEAPrepare}
\alias{GSEAPrepare}
\title{Title}
\usage{
GSEAPrepare(signature, Foldstatistic = "piFC", species)
}
\arguments{
\item{species}{}
}
\value{
Gene vector
}
\description{
Title
}
\examples{
GSEAPrepare( )
}
