% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanToMouse.R
\name{HumanToMouse}
\alias{HumanToMouse}
\title{Convert human gene symbols to mouse gene symbols}
\usage{
HumanToMouse(x)
}
\arguments{
\item{x}{character() Vector of human HGNC gene symbols.}
}
\value{
character() Unique vector of mapped mouse MGI symbols.
}
\description{
Maps a vector of human HGNC symbols to mouse MGI symbols using biomaRt and the
Ensembl datasets for Homo sapiens and Mus musculus.
}
\examples{
\dontrun{
  HumanToMouse(c("TNF", "IL6", "CD4"))
}
}
