% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGSEAalasDB.R
\name{MakeGSEAtlasDB}
\alias{MakeGSEAtlasDB}
\title{Title}
\usage{
MakeGSEAtlasDB(marker_df, p_val_adj = 0.05, avg_log2FC = 1)
}
\arguments{
\item{avg_log2FC}{}
}
\value{
Output annotation
}
\description{
Title
}
\examples{
MakeGSEAAtlasDB( )
}
