% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareTopBar.R
\name{PrepareTopBar}
\alias{PrepareTopBar}
\title{Prepare top annotations for ComplexHeatmap}
\usage{
PrepareTopBar(
  DiscreteAnnotations,
  DiscretePalettes,
  ContinuousAnnotations = NULL,
  ContinuousPalettes = NULL
)
}
\value{
A ComplexHeatmap::HeatmapAnnotation object suitable for use as top_annotation.
}
\description{
Prepare top annotations for ComplexHeatmap
}
\examples{
PrepareTopBar(DiscreteAnnotations,DiscretePalettes,ContinuousAnnotations=NULL,ContinuousPalettes=NULL)
}
