% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareTopBar.R
\name{PrepareTopBar}
\alias{PrepareTopBar}
\title{Title}
\usage{
PrepareTopBar(
  DiscreteAnnotations,
  DiscretePalettes,
  ContinuousAnnotations = NULL,
  ContinuousPalettes = NULL
)
}
\value{
Heatmap annotations
}
\description{
Title
}
\examples{
PrepareTopBar()
}
