% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareTopBar.R
\name{PrepareTopBar}
\alias{PrepareTopBar}
\title{Prepares the annotations for plotting with CompleHeatmap}
\usage{
PrepareTopBar(
  DiscreteAnnotations,
  DiscretePalettes,
  ContinuousAnnotations = NULL,
  ContinuousPalettes = NULL
)
}
\value{
Complex Heatmap annotations
}
\description{
Prepares the annotations for plotting with CompleHeatmap
}
\examples{
PrepareTopBar(DiscreteAnnotations,DiscretePalettes,ContinuousAnnotations=NULL,ContinuousPalettes=NULL)
}
