% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionScorePlot.R
\name{ProjectionScorePlot}
\alias{ProjectionScorePlot}
\title{Title}
\usage{
ProjectionScorePlot(ps.df)
}
\arguments{
\item{ps.df}{}
}
\value{
Projection Score Plot for quick variable selection
}
\description{
Title
}
\examples{
ProjectionScorePlot(ps.df)

}
