% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratMerge.R
\name{SeuratMerge}
\alias{SeuratMerge}
\title{Merge Multiple Seurat Objects}
\usage{
SeuratMerge(seurats)
}
\arguments{
\item{seurats}{A character vector of file paths to the Seurat objects to be merged.}
}
\value{
A merged Seurat object with joined layers.
}
\description{
This function reads multiple Seurat objects from a list of file paths, merges them into a single Seurat object, and joins the layers.
}
