% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAMPDeconvolPlot.R
\name{STAMPDeconvolPlot}
\alias{STAMPDeconvolPlot}
\title{Title}
\usage{
STAMPDeconvolPlot(ExprMat, Pheno, my_comparisons, Factor, Palette, Compartment)
}
\value{
PC score plot
}
\description{
Title
}
\examples{
GeneScorePlot(Signature = pcSig, Pheno = yf$samples,  my_comparisons = list(c("TumorDay1","SkinDay8") ,c("TumorDay3","SkinDay8"),c("TumorDay8","SkinDay8"),c("PoreDay1","SkinDay8"),c("PoreDay3","SkinDay8"),c("PoreDay8","SkinDay8")))
}
