% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoWrap.R
\name{VolcanoWrap}
\alias{VolcanoWrap}
\title{Wrapps EnhancedVolcanoPlot}
\usage{
VolcanoWrap(signature, Prefix, contraste, listofgenes = NULL)
}
\arguments{
\item{signature}{DE result.}

\item{Prefix}{String to append to title.}

\item{contraste}{Title of the plot.}

\item{listofgenes}{List of genes to highlight.}
}
\value{
Volcano plot
}
\description{
Wrapps EnhancedVolcanoPlot
}
\examples{
VolcanoWrap(signature,Prefix,contraste,listofgenes = NULL)

}
